BracketCheck class
two arrays track all bracket types for which the validator is good.
Uses a stack to track the most recent seen opening type bracket. If closing types are seen, they are checked against the top of the stack.
If a match is found, the stack is popped.
If a match is not found, or the stack is empty, validator returns false